mongodb_rbac/requirements.txt
coverage==7.6.1
dnspython==2.6.1
exceptiongroup==1.2.2
iniconfig==2.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
Pygments==2.18.0
pymongo==4.10.1
pytest==8.3.3
pytest-cov==5.0.0
python-dotenv==1.0.1
rich==13.9.4
tomli==2.0.2
typing_extensions==4.12.2
